cli/pyproject.toml
[project]
name = "pegasus-v2f"
version = "0.1.0"
description = "PEGASUS V2F — gene prioritization pipeline and database engine"
requires-python = ">=3.11"
dependencies = [
    "rich-click>=1.8",
    "duckdb>=1.0",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "httpx>=0.27",
    "rich>=13.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
sheets = ["gspread>=6.0"]
dev = ["pytest>=8.0"]

[project.scripts]
v2f = "pegasus_v2f.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pegasus_v2f"]

[tool.pytest.ini_options]
testpaths = ["tests"]
